Move Docker pull config factory to Shakedown.
Summary:
This is the start of moving helper mehtods from `tests/system/common.py`
to Shakedown.
